+++ src/main/java/com/birca/bircabackend/command/cafe/infrastructure/uploadcount/OcrRequestCountInterceptor.java
@@ -1,4 +1,4 @@
-package com.birca.bircabackend.command.cafe.validation;
+package com.birca.bircabackend.command.cafe.infrastructure.uploadcount;
 
 import com.birca.bircabackend.command.auth.application.token.TokenPayload;
 import com.birca.bircabackend.command.cafe.domain.OcrRequestHistoryRepository;
@@ -7,7 +7,6 @@ import com.birca.bircabackend.common.exception.InternalServerErrorCode;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import org.springframework.web.cors.CorsUtils;
 import org.springframework.web.method.HandlerMethod;
@@ -15,12 +14,10 @@ import org.springframework.web.servlet.HandlerInterceptor;
 
 import java.util.Optional;
 
-@Slf4j
 @Component
 @RequiredArgsConstructor
 public class OcrRequestCountInterceptor implements HandlerInterceptor {
 
-    private final OcrRequestHistoryRepository ocrRequestHistoryRepository;
     private final OcrRequestCountValidator ocrRequestCountValidator;
 
     @Override
@@ -36,8 +33,7 @@ public class OcrRequestCountInterceptor implements HandlerInterceptor {
 
     private boolean handleHandlerMethod(HttpServletRequest request) {
         TokenPayload payload = getPayload(request);
-        ocrRequestHistoryRepository.findUploadCountByOwnerId(payload.id())
-                .ifPresent(ocrRequestCountValidator::validateUploadCount);
+        ocrRequestCountValidator.validateUploadCount(payload.id());
         return true;
     }
